debug การลบ temp image กรณีที่ไม่พบไฟล์

+++ common/ajaxDeleteTempImage.php
@@ -1,7 +1,12 @@
 <?
-if(unlink($_REQUEST['imgPath'])) {
-	echo 'PASS';
+if(file_exists($_REQUEST['imgPath'])) {
+	if(unlink($_REQUEST['imgPath'])) {
+		echo 'PASS';
+	} else {
+		echo 'DELETE_TEMP_IMAGE_FAIL';
+	}
 } else {
-	echo 'DELETE_TEMP_IMAGE_FAIL';
+	echo "FILE_NOT_EXISTS";
 }
+
 ?>
